requirements.txt
Flask==2.1.2
flask-cors==3.1.1
numpy==1.23.5
PyPDF2==2.10.0
pandas==1.5.3
python-docx==0.8.11
sentence-transformers==2.2.2
faiss-cpu==1.7.3
streamlit==1.23.0
requests==2.28.1

